requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
astunparse==1.6.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
et_xmlfile==2.0.0
evdev==1.4.0
exceptiongroup==1.3.0
fastapi==0.115.12
fasteners==0.14.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.55.3
future==0.18.2
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.71.0
grpcio-status==1.71.0
gyp==0.1
h11==0.16.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
keras==3.9.2
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==9.3.0
namex==0.0.8
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.0
olefile==0.46
openai==1.86.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prettytable==3.16.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.5
pydantic_core==2.33.2
pyecharts==2.0.8
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-magic==0.4.24
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==14.0.0
rsa==4.9.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
setproctitle==1.2.2
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
systemd-python==234
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tqdm==4.67.1
trimesh==4.6.6
typing-inspection==0.4.1
typing_extensions==4.13.1
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.3
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wxPython==4.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
